+++ relay-server/src/actors/processor.rs
@@ -911,9 +911,23 @@ impl EnvelopeProcessor {
 
     /// Remove replays if the feature flag is not enabled
     fn process_replays(&self, state: &mut ProcessEnvelopeState) {
-        let replays_enabled = state.project_state.has_feature(Feature::Replays);
+        let replays_enabled: bool = state.project_state.has_feature(Feature::Replays);
         state.envelope.retain_items(|item| match item.ty() {
-            ItemType::ReplayEvent | ItemType::ReplayRecording => replays_enabled,
+            ItemType::ReplayEvent => {
+                if !replays_enabled {
+                    return false;
+                }
+
+                let parsed_replay = relay_replays::normalize_replay_event(&item.payload());
+                match parsed_replay {
+                    Ok(replay) => {
+                        item.set_payload(ContentType::Json, &replay[..]);
+                        true
+                    }
+                    Err(_) => false,
+                }
+            }
+            ItemType::ReplayRecording => replays_enabled,
             _ => true,
         });
     }